+++ src/main/java/Crisp/Command/AddCommand.java
@@ -1,3 +1,7 @@
+package Crisp.Command;
+import Crisp.Task.*;
+import Crisp.util.*;
+
 public class AddCommand extends Command {
     private Task task;
