+++ src/main/java/Crisp/Command/DeleteCommand.java
@@ -1,3 +1,6 @@
+package Crisp.Command;
+import Crisp.Task.*;
+import Crisp.util.*;
 public class DeleteCommand  extends Command {
     private final int index; // index of task to delete (1-based)
