+++ src/main/java/Crisp/Command/ExitCommand.java
@@ -1,3 +1,6 @@
+package Crisp.Command;
+import Crisp.Task.*;
+import Crisp.util.*;
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
