+++ src/main/java/Crisp/Task/Event.java
@@ -1,3 +1,5 @@
+package Crisp.Task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
