+++ src/main/java/Crisp/Task/Status.java
@@ -1,0 +1,16 @@
+package Crisp.Task;
+
+public enum Status {
+    NOT_DONE(" "),
+    DONE("X");
+
+    private final String icon;
+
+    Status(String icon) {
+        this.icon = icon;
+    }
+
+    public String getIcon() {
+        return icon;
+    }
+}
