+++ src/main/java/Crisp/Task/Task.java
@@ -1,3 +1,5 @@
+package Crisp.Task;
+
 public abstract class Task {
     protected String description;
     protected Status status;
