+++ src/main/java/Crisp/Task/TaskType.java
@@ -1,0 +1,17 @@
+package Crisp.Task;
+
+public enum TaskType {
+    TODO("T"),
+    DEADLINE("D"),
+    EVENT("E");
+
+    private final String icon;
+
+    TaskType(String icon) {
+        this.icon = icon;
+    }
+
+    public String getIcon() {
+        return icon;
+    }
+}
