+++ src/main/java/Crisp/Task/Todo.java
@@ -1,3 +1,5 @@
+package Crisp.Task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description, TaskType.TODO);
