+++ src/main/java/Crisp/util/Ui.java
@@ -1,3 +1,7 @@
+package Crisp.util;
+
+import Crisp.Task.Task;
+
 import java.util.Scanner;
 
 public class Ui {
